GitHub App: handle webhook events in task
